Change repository name to itfit.
