+++ README.md
@@ -3,16 +3,23 @@
 <!-- </h1><br> -->
 
 <h1 align="center">
+Itfit
+</h1>
+<h2 align="center">
 Interactive Fitter
-</h1><br>
+</h2><br>
 
-[![PyPI Downloads](https://img.shields.io/pypi/dm/interactive-fitter.svg?label=PyPI%20downloads)](
-https://pypi.org/project/interactive-fitter/)
+[![PyPI Downloads](https://img.shields.io/pypi/dm/itfit.svg?label=PyPI%20downloads)](
+https://pypi.org/project/itfit/)
 
 Simple, intuitive and interactive application to help fitting common functions to your data.
 
 <h1 align="center">
 <img src="docs/images/sample.gif" width="400" height="320" />
+</h1>
+
+<h1 align="center">
+<img src="docs/images/example.png" width="400" height="320" />
 </h1>
 
 # Instalation
