+++ docs/SUMMARY.md
@@ -11,10 +11,12 @@
             * [linear](reference/fitter/fit_functions/linear.md)
             * [quadratic](reference/fitter/fit_functions/quadratic.md)
         * [fitter_app](reference/fitter/fitter_app.md)
+        * [qt_fitter_app](reference/fitter/qt_fitter_app.md)
         * [utils](reference/fitter/utils/index.md)
             * [blit_manager](reference/fitter/utils/blit_manager.md)
             * [collection](reference/fitter/utils/collection.md)
             * [exponential](reference/fitter/utils/exponential.md)
+            * [fit_container](reference/fitter/utils/fit_container.md)
             * [gaussian](reference/fitter/utils/gaussian.md)
             * [line](reference/fitter/utils/line.md)
             * [point](reference/fitter/utils/point.md)
