+++ docs/reference/fitter/utils/fit_container.md
@@ -1,0 +1,1 @@
+:::fitter.utils.fit_container
