+++ itfit/__init__.py
@@ -1,16 +1,16 @@
 """TODO"""
 try:
-    __FITTER_IMPORTED__
+    __ITFIT_IMPORTED__
 except NameError:
-    __FITTER_IMPORTED__= False
+    __ITFIT_IMPORTED__= False
     
 __version__ = "0.0.4"
 
-if not __FITTER_IMPORTED__:
+if not __ITFIT_IMPORTED__:
     from . import data
     from . import fit_functions
     from . import data_selectors
     from .fitter_app import Fitter
     
     
-__FITTER_IMPORTED__ = True
+__ITFIT_IMPORTED__ = True
