+++ itfit/fit_functions/exponential.py
