+++ itfit/fit_functions/generic_fitter.py
