+++ itfit/fit_functions/quadratic.py
