+++ itfit/utils/blit_manager.py
