+++ itfit/utils/quadratic.py
