+++ mkdocs.yml
@@ -1,4 +1,4 @@
-site_name: "interactive-fitter"
+site_name: "itfit"
 site_description: "Simple and intuitive visual simple function fitter."
 # site_url: "https://mkdocstrings.github.io/"
 repo_url: "https://github.com/QuanticPony/interactive_fitter"
