+++ pyproject.toml
@@ -4,7 +4,7 @@ requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "interactive-fitter"
+name = "itfit"
 version = "0.0.4"
 requires-python = ">=3.8"
 
@@ -37,7 +37,7 @@ classifiers = [
 dependencies = ['numpy', 'matplotlib', 'scipy', 'mkdocstrings[python]>=0.18', 'mkdocs-material', 'bumpver', 'mkdocs', 'mkdocstrings[python]>=0.18', 'mkdocs-gen-files', 'mkdocs-material', 'mkdocs-literate-nav', 'Pygments>=2.12']
 
 [project.urls]
-"GitHub" = "https://github.com/QuanticPony/interactive_fitter"
+"GitHub" = "https://github.com/QuanticPony/itfit"
 
 [tool.bumpver]
 current_version = "0.0.4"
