+++ setup.py
@@ -1,4 +1,4 @@
 from setuptools import setup
 
 if __name__=='__main__':
-    setup(packages=["fitter"])
+    setup(packages=["itfit"])
